requirements.txt
alembic==1.3.3
aniso8601==8.0.0
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Mako==1.1.1
MarkupSafe==1.1.1
pandas==0.25.3
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
wincertstore==0.2
